refactor (HTML): changed a couple of things in the management.html file - changed html parts to a more meaningful naming
