+++ platformio.ini
@@ -9,17 +9,17 @@
 ; https://docs.platformio.org/page/projectconf.html
 
 ;[env:nodemcu-32s]
-[env:lolin32]
+[env:az-delivery-devkit-v4]
 platform = espressif32
-;board = nodemcu-32s
-board = lolin32
+board = az-delivery-devkit-v4
 framework = arduino
 monitor_speed = 115200
 board_build.partitions = no_ota.csv
-;board_build.partitions = min_spiffs.csv
-
 upload_port = /dev/cu.SLAB_USBtoUART
 monitor_port = /dev/cu.SLAB_USBtoUART
+build_flags = -O2
+build_unflags = -Os
+
 
 lib_deps =
  https://github.com/schreibfaul1/ESP32-audioI2S.git
